Released OAFC (version 5.0)
Update Notes:

The script was slightly changed for more efficiency, and since the devs added the "Auto" button (huge W for that), it now simply cracks the button for it to be always active.
Now you only have to fetch the link (execute the script) and click that "Auto" button when you need some auto merging. As simple as that!
And with those improvements, the script now does not conflict with any UI, so you're safe to use the app however you want even when the script is executed!